fix: rename optionSearch
